hellworld/helloworld_test.go
package main

import "testing"

func TestHello(t *testing.T) {
	t.Run("saying hello to people", func(t *testing.T) {
		got := Hello("Abhishek", "")
		want := "Hello, Abhishek"
		assertCorrectMessage(t, got, want)
	})

	t.Run("saying hello in spanish", func(t *testing.T) {
		got := Hello("Abhishek", spanish)
		want := spanishHelloPrefix + "Abhishek"
		assertCorrectMessage(t, got, want)
	})

	t.Run("saying hello world", func(t *testing.T) {
		got := Hello("", "")
		want := "Hello, World"
		assertCorrectMessage(t, got, want)
	})
}

func assertCorrectMessage(t testing.TB, got, want string) {
	t.Helper()
	if got != want {
		t.Errorf("got %q want %q.", got, want)
	}
}
